@@ -65,11 +65,11 @@
     "rollup-watch": "^4.3.1",
     "selenium-server": "^2.53.1",
     "todomvc-app-css": "^2.1.0",
-    "typescript": "^2.1.5",
+    "typescript": "^2.5.2",
     "uglify-js": "^3.1.2",
-    "vue": "^2.4.3",
+    "vue": "git+https://github.com/vuejs/vue.git#96b97448118de0939bf5f77c9b74cf1613a5a107",
     "vue-loader": "^13.0.5",
-    "vue-template-compiler": "^2.4.3",
+    "vue-template-compiler": "^2.4.4",
     "webpack": "^3.6.0",
     "webpack-dev-middleware": "^1.10.0",
     "webpack-hot-middleware": "^2.16.1"
