@@ -1,4 +1,4 @@
-import _Vue = require("vue");
+import _Vue from "vue";
 import { WatchOptions } from "vue";
 
 // augment typings of Vue.js
